+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v3.11.0-dev
 
 - Enhancements
+    - Default upstream database version is now `6.4.0`
     - Upstream improvements for ARM platform detection have been integrated
     - Upstream improvements for Blink engine version detection have been integrated
     - Upstream improvements for browser version detection have been integrated